+++ lib/util/web/src/telegram_web_app_dummy.dart
@@ -8,4 +8,10 @@ class Telegram {
   String? getTgUserId() => null;
 
   void hapticFeedback() {}
+
+  Future<String?> showScanQrPopup() async => null;
+
+  String get platform => '';
+
+  String get version => '';
 }
