+++ lib/util/web/telegram_web_app.dart
@@ -1,41 +1,2 @@
-// ignore: avoid_web_libraries_in_flutter
-import 'dart:js' as js;
-
-import '../logger.dart';
-
-class Telegram {
-  Telegram._internal() {
-    final telegram = js.context['Telegram'] as js.JsObject;
-    webApp = telegram['WebApp'] as js.JsObject;
-  }
-
-  static Telegram instance = Telegram._internal();
-
-  late js.JsObject webApp;
-
-  String? getTgInitData() {
-    final initData = webApp['initData'];
-    if (initData == null) {
-      return null;
-    }
-    return initData as String;
-  }
-
-  String? getTgUserId() {
-    final initDataUnsafe = webApp['initDataUnsafe'] as js.JsObject;
-    final user = initDataUnsafe['user'];
-    if (user == null) {
-      return null;
-    }
-    return (user as js.JsObject)['id'] as String;
-  }
-
-  void hapticFeedback() {
-    try {
-      (webApp['HapticFeedback'] as js.JsObject)
-          .callMethod('impactOccurred', ['light']);
-    } catch (error, stacktrace) {
-      e('hapticFeedback error $error, $stacktrace');
-    }
-  }
-}
+export 'src/telegram_web_app.dart'
+    if (dart.library.io) 'src/telegram_web_app_dummy.dart';
